Created .hpp & .cpp files for the ExampleConsumeCb & MyHashPartitionerCb classes.
