+++ src/exampleConsumeCb.cpp
@@ -1,0 +1,5 @@
+#include "exampleConsumeCb.h"
+
+void ExampleConsumeCb::consume_cb(RdKafka::Message &msg, void *opaque) {
+    //msg_consume(&msg, opaque);
+}
